requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-genai==0.6.0
pytest==8.3.3
httpx==0.27.2
